move generate functions to mocs for trading e2e to avoid violating boundary rules
