+++ apps/trading-e2e/src/support/mocks/generate-market.ts
@@ -1,6 +1,15 @@
 import merge from 'lodash/merge';
 import type { PartialDeep } from 'type-fest';
-import type { Market } from '../__generated__/Market';
+
+export interface Market_market {
+  __typename: 'Market';
+  id: string;
+  name: string;
+}
+
+export interface Market {
+  market: Market_market | null;
+}
 
 export const generateMarket = (override?: PartialDeep<Market>): Market => {
   const defaultResult = {
