--- libs/chart/src/__tests__/index.ts
@@ -1,2 +1,0 @@
-export * from './generate-candles';
-export * from './generate-chart';
