+++ libs/chart/src/index.ts
@@ -1,1 +1,5 @@
 export * from './lib/trading-chart';
+export * from './lib/__generated__/Candles';
+export * from './lib/__generated__/CandleFields';
+export * from './lib/__generated__/CandlesSub';
+export * from './lib/__generated__/Chart';
