--- libs/deal-ticket/src/__tests__/index.ts
@@ -1,1 +1,0 @@
-export * from './generate-deal-ticket-query';
