+++ libs/deal-ticket/src/index.ts
@@ -1,3 +1,5 @@
 export * from './deal-ticket';
 export * from './use-order-state';
 export * from './deal-ticket-container';
+export * from './__generated__/DealTicketQuery';
+export * from './__generated__/OrderEvent';
