+++ libs/market-list/src/index.ts
@@ -1,2 +1,5 @@
 export * from './lib/market-list-table';
 export * from './lib/markets-container';
+export * from './lib/__generated__/Markets';
+export * from './lib/__generated__/MarketDataFields';
+export * from './lib/__generated__/MarketDataSub';
