+++ libs/trades/src/index.ts
@@ -1,1 +1,4 @@
 export * from './lib/trades-container';
+export * from './lib/__generated__/TradeFields';
+export * from './lib/__generated__/Trades';
+export * from './lib/__generated__/TradesSub';
